Fixes to WireGuard Connection Status
Modified code to correctly detect when a WireGuard interface is up (connected) or down (disconnected) in addition to being not enabled. Since a WireGuard connection is essentially stateless, the new code improves on the detection method to provide the correct status so the user can select the desired interface to run a speed test.

NOTE: This fix was provided by @ExtremeFiretop.